+++ packages/psychart/src/index.ts
@@ -5,5 +5,5 @@
  * ![NPM Downloads](https://img.shields.io/npm/d18m/psychart)
  * ![NPM Last Update](https://img.shields.io/npm/last-update/psychart)
  */
-export * from './psychart';
-export * from './types';
+export * from './psychart/psychart';
+export * from './psychart/types';
