+++ packages/psychart/src/psychart/defaults.ts
